move eval related functions to eval_utils.py
